src/backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Machine Learning and AI dependencies
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
transformers==4.35.2
sentence-transformers==2.2.2
peft==0.6.2
accelerate==0.24.1
safetensors==0.4.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Audio processing
librosa==0.10.1
soundfile==0.12.1
whisper==1.1.10

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
sumy==0.11.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client and async support
httpx==0.25.2
aiofiles==23.2.1
aiohttp==3.9.1

# Database (optional, for future use)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching and performance
redis==5.0.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.12.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# GPU support (optional)
# Uncomment if you have CUDA-capable GPU
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio==0.16.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.16.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
typing-extensions==4.8.0
packaging==23.2

# CORS and middleware
python-cors==1.7.0

# File handling
chardet==5.2.0

# Memory profiling (development)
memory-profiler==0.61.0
psutil==5.9.6

# API documentation
markdown==3.5.1

# Async task queue (optional)
celery==5.3.4

# Model serving optimization
onnx==1.15.0
onnxruntime==1.16.3

# Quantization and optimization
bitsandbytes==0.41.3

# Progress bars and CLI
tqdm==4.66.1
rich==13.7.0

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Metrics and evaluation
scikit-learn==1.3.2
scipy==1.11.4

# Time zone handling
pytz==2023.3

# JSON handling
orjson==3.9.10

# WebSocket support
websockets==12.0

# Health checks and monitoring
healthcheck==1.3.3

# Rate limiting
slowapi==0.1.9

# Background tasks
backgroundtasks==1.0.0

# File upload handling
streaming-form-data==1.13.0

# Model versioning
huggingface-hub==0.19.4

# Distributed computing (optional)
ray==2.8.0

# Profiling
py-spy==0.3.14

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0